vampirism

+++ Assets/Scripts/UI/Health/HealthBar.cs
@@ -10,7 +10,7 @@ public class HealthBar : UIHealthBase
         _slider = GetComponent<Slider>();
     }
 
-    public override void Set(int current, int max)
+    public override void Set(float current, float max)
     {
 
         _slider.maxValue = max;
