+++ Assets/Scripts/UI/Health/UIHealthBase.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public abstract class UIHealthBase : MonoBehaviour
 {
-    public abstract void Set(int current, int max);
+    public abstract void Set(float current, float max);
 }
